# apps/shared/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/events": {
      "post": {
        "description": "Create a new photobooth event with pricing information",
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEventResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create event",
        "tags": [
          "Events"
        ]
      },
      "get": {
        "description": "Retrieve all events created by the authenticated user",
        "operationId": "EventsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all events",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}": {
      "get": {
        "description": "Retrieve a specific event by its UUID",
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get event by ID",
        "tags": [
          "Events"
        ]
      },
      "patch": {
        "description": "Update specific fields of an event",
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Partially update event",
        "tags": [
          "Events"
        ]
      },
      "put": {
        "description": "Completely replace an event with new data",
        "operationId": "EventsController_replace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event replaced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Replace event",
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "description": "Permanently delete an event and all associated data",
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete event",
        "tags": [
          "Events"
        ]
      }
    },
...
    "/api/analytics/total": {
      "get": {
        "description": "Get aggregated analytics across all events or a specific event including revenue, session times, and print statistics",
        "operationId": "AnalyticsController_getTotalAnalytics",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Event ID to filter analytics for a specific event",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for analytics range (ISO string)",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for analytics range (ISO string)",
            "schema": {
              "example": "2024-12-31T23:59:59.999Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Total analytics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalAnalyticsDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date range or event ID provided"
          },
          "401": {
            "description": "Unauthorized - invalid or missing JWT token"
          },
          "404": {
            "description": "Event not found (when eventId is provided)"
          },
          "500": {
            "description": "Internal server error while calculating analytics"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get total analytics",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/total/monthly-trends": {
      "get": {
        "description": "Get aggregated analytics with trend calculations comparing current month to previous month",
        "operationId": "AnalyticsController_getTotalAnalyticsWithMonthlyTrends",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Event ID to filter analytics for a specific event",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Total analytics with monthly trends retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendTotalAnalyticsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing JWT token"
          },
          "404": {
            "description": "Event not found (when eventId is provided)"
          },
          "500": {
            "description": "Internal server error while calculating analytics"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get total analytics with monthly trends",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/events": {
      "get": {
        "description": "Get analytics data for each event including earnings, session times, and print counts. Use eventId to filter for a specific event.",
        "operationId": "AnalyticsController_getEventAnalytics",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Event ID to filter analytics for a specific event",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for analytics range (ISO string)",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for analytics range (ISO string)",
            "schema": {
              "example": "2024-12-31T23:59:59.999Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event analytics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventAnalyticsDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid date range or event ID provided"
          },
          "401": {
            "description": "Unauthorized - invalid or missing JWT token"
          },
          "404": {
            "description": "Event not found (when eventId is provided)"
          },
          "500": {
            "description": "Internal server error while calculating analytics"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get per-event analytics",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/events/{eventId}": {
      "get": {
        "description": "Get detailed analytics data for a specific event including earnings, session times, and print counts",
        "operationId": "AnalyticsController_getSingleEventAnalytics",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for analytics range (ISO string)",
            "schema": {
              "example": "2024-12-31T23:59:59.999Z",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for analytics range (ISO string)",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event analytics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventAnalyticsDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date range provided"
          },
          "401": {
            "description": "Unauthorized - invalid or missing JWT token"
          },
          "404": {
            "description": "Event not found or inactive"
          },
          "500": {
            "description": "Internal server error while calculating analytics"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get analytics for a specific event",
        "tags": [
          "Analytics"
        ]
      }
    }
  },
  "info": {
    "title": "Easy Picsy - Photobooth Payment API",
...
    "version": "1.0",
    "contact": {
      "name": "Easy Picsy",
      "url": "https://github.com/your-org/easy-picsy",
      "email": "support@easypicsy.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter your Supabase JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Event name",
            "example": "Birthday Party Photobooth",
            "minLength": 1,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Event description",
            "example": "Capture memories at John's 25th birthday celebration"
          },
          "price": {
            "type": "number",
            "description": "Price per photobooth session",
            "example": 50,
            "minimum": 20,
            "format": "decimal"
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "PHP",
            "default": "PHP",
            "minLength": 3,
            "maxLength": 3
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the event is active",
            "example": true,
            "default": true
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
...
      "CreateEventResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Event unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Event name",
            "example": "Birthday Party Photobooth"
          },
          "description": {
            "type": "string",
            "description": "Event description",
            "example": "Capture memories at John's 25th birthday celebration"
          },
          "price": {
            "type": "string",
            "description": "Price per photobooth session",
            "example": "50.00",
            "format": "decimal"
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "PHP"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the event is active",
            "example": true
          },
          "createdBy": {
            "type": "string",
            "description": "User ID who created the event",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Event creation timestamp",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Event last update timestamp",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "qrCode": {
            "description": "Generated QR code for the event (if successful)",
            "allOf": [
              {
                "$ref": "#/components/schemas/CurrentQrCodeResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "currency",
          "isActive",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "EventResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Event unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Event name",
            "example": "Birthday Party Photobooth"
          },
          "description": {
            "type": "string",
            "description": "Event description",
            "example": "Capture memories at John's 25th birthday celebration"
          },
          "price": {
            "type": "string",
            "description": "Price per photobooth session",
            "example": "50.00",
            "format": "decimal"
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "PHP"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the event is active",
            "example": true
          },
          "createdBy": {
            "type": "string",
            "description": "User ID who created the event",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Event creation timestamp",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Event last update timestamp",
            "example": "2023-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "currency",
          "isActive",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Event name",
            "example": "Birthday Party Photobooth",
            "minLength": 1,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Event description",
            "example": "Capture memories at John's 25th birthday celebration"
          },
          "price": {
            "type": "number",
            "description": "Price per photobooth session",
            "example": 50,
            "minimum": 20,
            "format": "decimal"
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "PHP",
            "default": "PHP",
            "minLength": 3,
            "maxLength": 3
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the event is active",
            "example": true,
            "default": true
          }
        }
      },
      "EventDeleteResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Deletion confirmation message",
            "example": "Event deleted successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "TotalAnalyticsDto": {
        "type": "object",
        "properties": {
          "totalNetRevenue": {
            "type": "number",
            "description": "Total net revenue from all completed payments",
            "example": 15750
          },
          "totalWithdrawableRevenue": {
            "type": "number",
            "description": "Total withdrawable/cashout-able revenue",
            "example": 15750
          },
          "averageSessionTime": {
            "type": "number",
            "description": "Average session time in seconds",
            "example": 180
          },
          "totalPrints": {
            "description": "Print statistics breakdown",
            "allOf": [
              {
                "$ref": "#/components/schemas/TotalPrintAnalyticsDto"
              }
            ]
          }
        },
        "required": [
          "totalNetRevenue",
          "totalWithdrawableRevenue",
          "averageSessionTime",
          "totalPrints"
        ]
      },
...
      "TrendTotalAnalyticsDto": {
        "type": "object",
        "properties": {
          "totalNetRevenue": {
            "type": "number",
            "description": "Current period total net revenue",
            "example": 15750
          },
          "totalWithdrawableRevenue": {
            "type": "number",
            "description": "Current period total withdrawable revenue",
            "example": 15750
          },
          "averageSessionTime": {
            "type": "number",
            "description": "Current period average session time in seconds",
            "example": 180
          },
          "totalPrints": {
            "description": "Current period print statistics with trends",
            "allOf": [
              {
                "$ref": "#/components/schemas/TrendPrintAnalyticsDto"
              }
            ]
          },
          "totalNetRevenueTrend": {
            "description": "Trend data for total net revenue",
            "allOf": [
              {
                "$ref": "#/components/schemas/TrendDataDto"
              }
            ]
          },
          "averageSessionTimeTrend": {
            "description": "Trend data for average session time",
            "allOf": [
              {
                "$ref": "#/components/schemas/TrendDataDto"
              }
            ]
          }
        },
        "required": [
          "totalNetRevenue",
          "totalWithdrawableRevenue",
          "averageSessionTime",
          "totalPrints",
          "totalNetRevenueTrend",
          "averageSessionTimeTrend"
        ]
      },
      "EventAnalyticsDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "description": "Event ID",
            "example": "event-uuid-123"
          },
          "eventName": {
            "type": "string",
            "description": "Event name",
            "example": "Wedding Photo Booth"
          },
          "runningEarnings": {
            "type": "number",
            "description": "Running earnings for this event",
            "example": 2500
          },
          "sessionAverageTime": {
            "type": "number",
            "description": "Session average time for this event in seconds",
            "example": 165
          },
          "numberOfPrints": {
            "type": "number",
            "description": "Number of prints for this event",
            "example": 24
          },
          "numberOfReprints": {
            "type": "number",
            "description": "Number of reprints for this event",
            "example": 7
          }
        },
        "required": [
          "eventId",
          "eventName",
          "runningEarnings",
          "sessionAverageTime",
          "numberOfPrints",
          "numberOfReprints"
        ]
      }
    }
  }
}